SDL2 Compiling in Windows
This is not yet cross platform, I need to do some work on the header
file and the makefile to get it to do this. However it is working and
working is good. This took a lot longer than I wanted it too. I went
ahead and included the SDL library with the project because otherwise
you'd have to make the correction to their header files that I did.
Because SDL 2.0.3 is released in a broken state. Don't worry I got you
bro... Its fixed and all it currently does is open a window and then
close after 3000 Miliseconds. But its something.