So first round of code corrections
So Juan made a version of the code that made it so it closed on x press
based off of John's code. But he didnt use Johns code so he overwrote
John's code. So i merged the code of the two but john's code mostly took
a back seat and is most of the commented out code i tried to show whose
code was which in comments. Then the line on 142 SDL_FillRect by john
breaks the program idk why. So its commented out.