Game Loops adjustments
Moved David's work to a .h file. Not enough for a full class yet but
easy to expand on the file. Mostly for organization for now.

Moved code that changes screen to the draw() function.

Moved OnEvent function to the update() function. makes more sense there
since it is updating game based on events.