Object and Physics Classes semi-functional
Still a work in progress, but my laptop was giving me issues and we need
to turn in a report today, so I figured I'd go ahead and commit while my
internet's functioning. Not a whole lot in terms of usability yet, but
y'all can look it over if you want to see how I'm going about certain
things. Both classes are in the Physics folder right now but will be
moved to separate folders when I get around to making a proper makefile.
I've went ahead and included my test file and the makefile associated
with it in the folder as well in case y'all want to play around with
them without any hassle, let me know if I should leave them out of the
next commit.