Gamepad.h
So i made a class for game pads that i can use. It still has some issues
due to the fact that mingw doesnt want to use the real XInput lib. At
the moment it is using a fake version. Due to that one function is
broken and it is commented out at the moment.  but the code is compling
and working at this time