Frame timing and some Physics
Updated gameroot to tell Physics how long the previous frame took; the
approach I used requires c++11, so I updated the makefile to compile for
that, and also put the Physics and Object .cpp files to be compiled; we
may want to make .o files so the entire thing doesn't have to be
recompiled each time any file is changed, but it won't make a large
difference. For Physics I've made progress with collision detection and
handling, but there's still a lot to do and is far from implementable. I
created a header file for Player and NonPlayerCharacters which inherit
from Object, but they are subject to change and don't have any
additional functionality yet.