Physics changes and updated gameroot.cpp
Collisions are acting pretty strangely at the moment, but I'll be
working on them tonight. Right now each object's movement has to be
updated manually in gameroot::update(), but I'll be adding a single
function that handles all movement appropriately that just needs one
line in the appropriate place in gameroot::update(). I'll construct the
objects for the level and get the platform moving as desired in a bit,
but wanted a commit with only the necessary components in place.