Timer updates
I've updated the timer to be a member of the gameroot class so all
functions have access to it if they need and moved some of the logic to
more appropriate places. With this method, there's a C++11 function
called std::this_thread::sleep_for (or sleep_until) that will allow for
a cross-platform and very precise waiting to control our frames per
second, but minGW doesn't support threads properly so it's not currently
implemented. I've left John's timers and waiting code in for now, but
it's going to be replaced with the new timers in the future.