Behavior readying
The Behavior class is pretty much set up and ready to have some default
AI implemented, but I'll need the new Physics and some Statistics before
there's any actually intelligent behavior for Objects. Note that there's
a chance I'll be making a class that inherits from Objects that will
have a Behavior member and removing the current member from Object; I'm
not sure on a name, but it should encompass all non-player characters
(enemies, allies, neutral characters), traps, items/powerups in the
world (But not the inventory), and destructible and interactable objects
(barrels, levers, etc.). There may be further subclasses of this class,
but at the moment I'm not even certain if it will be necessary as Player
may use Behavior as well, making it a pretty unnecessary distinction.