Size trimming
Implemented a method of shaving off 2 bytes for each of the members
inside of an instance of Object which require access to other members'
data (Behavior, Sprite/Animation, and Input). So, Behavior is now the
same size as an int- just 4 bytes- and is set up to be able to do quite
a bit. I am very happy with this, but am going to have to be adding a
few more bytes for extra functionality later on.