Even more trimming
Managed to take out the index from within Object and just use the one
that's already need inside of Behavior, so each Object only takes the
size of its components plus 4 bytes for the pointer in the static
vector. I have it set up to be able to do a similar thing for Player and
input, though there will be another 4 bytes since it's useful to have a
pointer in both the Object and Player pointer vectors.