Began work on Attributes
I've made a generic list of attributes that many types of games could
make use of and will be defining rules of how they interact, which users
will be able to modify to a certain extent. I've also tried to ensure
the attributes were easy to expand upon without causing issues with
predefined attributes. I figured this approach would be a little easier
for someone trying to use the engine to modify to fit their game rather
than forcing them to define all attributes and rules themselves. I tried
to be as non-genre specific as I could, and with a bit of creative
leeway on what the names represent within the game, I think it should be
able to handle most RPGs, platforming games, shooters, exploration
games, beat-em-ups, and ship-fighter games, and potentially stealth
games, though for some genres it's taking up a lot more memory than
necessary.

I'm going to be starting some work on an Inventory and Equipment class
pretty soon as well.