Physics, Attributes, Equipment
Object now has some support for Box2D's physics, but I haven't tested
much of it out. The makefile is now ultra massive since there were a
million compiler errors when trying to use any Box2D stuff before and
I'm not the best with messing with compilers, so I made a quick fix that
(presumably) works. Attributes and Equipment have limited functionality,
but you should be able to see more or less what I'm going for.
Currently, it's best to use the Object constructor that has the physics
information passed as parameters; I'll update the other ones later, but
the physics member will be pointing to a random memory location for most
of the others at the moment. Let me know if there are any problems you
notice aside from that though.