Makefile edits and moving Object to a folder
Just updating the Makefile to be a little bit better and reducing the
number of redundant includes in a few pieces of code, still could use
some improvements. I also moved all Object components and future derived
class header and cpp files to their own folder to clear up the workspace
a bit.