Error handling, commenting, rearranging
The Object constructor for Physics should be ready to go for
single-fixture rectangles; bodies with multiple fixtures and other
shapes will have to wait. All Object constructors and destructors should
execute without any issues, though admittedly I didn't do an extensive
test on the Physics yet. I may remove the error handling for when a
Box2D World hasn't been prepared as we'll always be providing one, but
for now it'll stay.