Attributes and Equipment mostly finished
Attributes, Equipment, and Items are nearing their final stages; you can
now define rules of how much a selection of attributes affect other
attributes, store innate attributes, can equip any number of items to an
Object, and can calculate the 'effective attribute' of all types of
attributes. For a few variables, this is the same as the value stored in
Attributes, but I may provide a function with a similar naming standard
to avoid possible confusion. Regeneration effects will be up and running
once the proper time step is passed to them, but for now they assume
it's running at 60 frames per second and shouldn't go over the effective
maximum of their respective attributes. Feel free to move any private
members of Object to public if you need them for the demo.