Menu class start
there is now a functional Menu that opens, allowing the user to pick if
they want to play a level, or edit a level. If engine is opened with a
click, double click to make choice. If opened with enter, just click.

Eventually will be put on loop so newly created levels can be tested.