Status_Effects implemented + Attributes updates
Status_Effects are now able to be applied to Objects; currently, you
must create a Status_Effect instance and change the modifiers
appropriately, then apply the Status_Effect to the Status_Effects member
of the Object with a call to "apply_Status_Effect(Status_Effect&)",
which will append the Status_Effect to a vector. To remove it, call
'remove_Status_Effect(std::uint8_t index_Of_Status_Effect)"; note that
currently the length of the vector containing pointers to active status
effects will be unaffected by the call and simply changes the pointer at
the index to nullptr, but will be changing in the future. Also note that
a Status_Effect can only be applied to one Object at a time at the
moment; support one Status_Effect applying to multiple Objects will be
added in the future. Items and Equipment work similarly, but will remain
with the limit of only being able to be equipped by one Object.
Effective Attributes should be functioning correctly and generally
follow the formula "(Innate Attribute + Equipment Modifiers +
Status_Effects Modifier ) * Status_Effects Multiplier". The base
multiplier is 1, and can go up or down from there depending on active
Status_Effects. Support for a leveling system will be implemented in the
future, which will give a variable amount of points to distribute
between primary attributes (Strength, Dexterity, Endurance,
Intelligence, and Perception) which will also have a variable cost to
increase. Also fixed a lot of 'const' declarations being in the wrong
spot.