Changed Map Editor to work with Level class
Note that I removed the check on the input string due to my editor
saving the files with some beginning of file character, but the output
files generated by using the editor shouldn't have them; feel free to
change it to how you see fit. To make a level just call the Level
constructor with the full file path passed as the only parameter, right
now if you exceed the maximum Objects (Currently set to 10), it will
just make 10 Objects, but I'll have a back-up Vector later on that will
allow some heap-allocated Objects in addition to the stack-allocated
ones. We'll also have to do some restructuring of code for this to make
any difference at the moment, but that will come later, if we get there.
Let me know if there's any problems.