Added Comments And Collision is odd.
I added comments to the input values in the map editor. I figured they
needed to be in both places.

I also added new platforms and made two of the platforms static by hand
so we could test collision.

And something is off by a long shot. In addition to that Objects are
not keeping their correct sizes on import to the level class.