Output issues in Map Editor
If you need to half the size of the platforms for physics do it on the
physics side. Otherwise you’ll slowly ruin the map files. Each time the
file is loaded it loads in 1 to 1 and then outputs as 1 to 1/2 causing
the objects in the level to slowly shrink. I removed this code and will
test to see how if at all it breaks the physics

The Physics seem to be just fine and broken still the way they were
before.

http://cdn.meme.am/instances/60338110.jpg