Quick fixes
There's definitely some issues with the draw function as it's currently
assuming the Object in level matches up with the Object in the pointer
vector, but it should look a bit better now. Should look a little less
buggy, but there's probably some issues with all of the Objects starting
inside one another or perhaps some not being rendered properly.