More quick fixes
The renderer was assuming the Objects to be 200 x 30; added a temporary
fix to the issue. It should display properly now so long as all Objects
are 200 x 30, but the Physics works for anything as far as I'm aware,
it's just going to be hard to tell what's going on until an actual
function handling drawing is in place. Also forgot to mention I had
changed the Level class to half the dimensions for the X and Y on the
last commit along with changing the update function on level to just use
one call to Update since it's the static function and would be stepping
the world once for each Object instead of just once each frame.