Matching renderer with map file
The current implementation to draw to the screen assumes Objects are 200
x 30, so it will draw them that size regardless of their actual size; I
believe the previous merge reverted the changes I had made to the map
file. When I get a chance to work on the project I'll make a proper
rendering function for Objects that uses Marcelo's Animation class, but
for now if you're trying to test Physics change the renderer and map
file to match the sizes or know that what's displayed doesn't correspond
to what's actually happening for Physics.