Drawing Objects from editor
Not completely finished, kind of bypasses Animation and just uses its
texture for now, but it's in a working state; needs to implement a
camera which shouldn't be difficult, but I ran out of time to work on it
for now. It'll read from the map file like it did before, but assigns
the image located in the path string to Animation's texture and displays
based on the width and height at creation, and current location based on
Physics; it's very dependant on being a rectangle right now though, so
some changes will have to be made once we implement other shapes.