Constant FPS, Smooth Animations, Tweaks to Player
I plan on making more changes but the title tells it all.

I changed the way the animation class worked and then spent time
working on the FPS timing to make it more exact with the help of
Andrew.

We took the update out of event to prevent the random update speed
issue.