Camera Zooms and follows Hero
Updated a few things to where there's some zoom functionality using a
static float from Object, but I haven't implemented a way to change it
without adjusting it and recompiling yet. The level should be zoomed in
twice as much with the current settings, and in Camera the width is
different from the window which is different from the standard
resolution (1282 vs 1024 vs 1080); I left the values alone for now in
case anyone's parts relied on that, but if not, we should probably have
that at 1080. All it means for when you're playing is that the camera
won't follow until you're a little passed the middle. The camera also
doesn't take into account the player's size, and uses the top left
corner from Animation; I can change that if needed. Animation and
Objects now uses a slightly modified render call that let's me change
the width and height for the zoom.

Current problems that I'm aware of at the moment are some slight pixel
overlaps on the Player standing on a platform or moving left into a
platform which is due to hard-coded values for Player physics not
corresponding properly with the sprite, Objects reaching Box2D's maximum
speed rather quickly (~120 meters per second, varies depending on frame
rate), and the aforementioned camera center; once levels are using
smaller values for Physics the max speed shouldn't be a problem we'll
need to address.

Let me know if you notice any other errors.